% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_finctions.R
\name{asinInverseTransform}
\alias{asinInverseTransform}
\title{Inverse arc-sine transformation}
\usage{
asinInverseTransform(p)
}
\arguments{
\item{p}{numeric vector}
}
\value{
numeric vector representing inverse of arc-sine transformation of p
}
\description{
Inverse arc-sine transformation
}
\examples{
asinInverseTransform(fitted(reproduce_lmm_CORRECTNESS()))
}
