% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_finctions.R
\name{asinTransform}
\alias{asinTransform}
\title{arc-sine transformation}
\usage{
asinTransform(p)
}
\arguments{
\item{p}{numeric vector}
}
\value{
arc-sine transformation of p
}
\description{
arc-sine transformation
}
\examples{
asinTransform(reproducerTaskGra::data_experiment$COR)
}
