% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create box plot with facet}
\usage{
create_boxplot(data_plot, x_variable, y_variable, facet_variable,
  ifStat = FALSE)
}
\arguments{
\item{data_plot}{a dataframe containing the data for plot}

\item{x_variable}{name of the factor for x-axis}

\item{y_variable}{name of the variable for y-axis}

\item{facet_variable}{name of the factor for creating facets}

\item{ifStat}{Should the mean and mean values added to the plot}
}
\value{
a ggplot object containing the violin plot
}
\description{
Create box plot with facet
}
\examples{
create_boxplot(reproducerTaskGra::data_experiment, "TASK_GRA", "COR", "TASK")
}
