% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_emmeans_profileplot}
\alias{create_emmeans_profileplot}
\title{Creates the profile plots from the estimated marginal means}
\usage{
create_emmeans_profileplot(data_plot, x_variable = "TASK_GRA",
  y_variable = "emmean", group_variable = "TASK")
}
\arguments{
\item{data_plot}{a data_frame containing the estimated marginal means for each level of x_variable and group_variable}

\item{x_variable}{column name in data_plot that contains the factor for x-axis, e.g. "TASK_GRA"}

\item{y_variable}{column name in data_plot that contains the estimated marginal means}

\item{group_variable}{column name in data_plot that contains the factor for which seperate profile plots will be shown for each level}
}
\value{
a ggplot object containing the profile plot
}
\description{
Creates the profile plots from the estimated marginal means
}
\examples{
emmeans_CORRECTNESS <- reproduce_emmeans_CORRECTNESS()
create_emmeans_profileplot(data_plot= emmeans_CORRECTNESS,
x_variable="TASK_GRA", y_variable="emmean", group_variable="TASK")
}
