% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_violin_plot}
\alias{create_violin_plot}
\title{Create violin plot}
\usage{
create_violin_plot(data_plot, x_variable, y_variable, ifStat = FALSE)
}
\arguments{
\item{data_plot}{a dataframe containing the data for plot}

\item{x_variable}{name of the variable for x-axis}

\item{y_variable}{name of the variable for y-axis}

\item{ifStat}{Should the mean and mean values added to the plot}
}
\value{
a ggplot object containing the violin plot
}
\description{
Create violin plot
}
\examples{
create_violin_plot(get_data(), "TASK_GRA", "COR")
create_violin_plot(reproducerTaskGra::data_experiment, "TASK", "TDD_CONF_SELF")
}
