% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotViolin}
\alias{plotViolin}
\title{Create violin plot}
\usage{
plotViolin(data_plot, x_variable, y_variable, ifStat = FALSE)
}
\arguments{
\item{data_plot}{a dataframe containing the data for plot}

\item{x_variable}{name of the variable for x-axis}

\item{y_variable}{name of the variable for y-axis}

\item{ifStat}{Should the mean and mean values added to the plot}
}
\value{
a ggplot object
}
\description{
Create violin plot
}
\examples{
plotViolin(get_data(), "TASK_GRA", "COR")
}
