% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_barplot_participants}
\alias{reproduce_barplot_participants}
\title{Reproduce bar plot for participants' demographics data}
\usage{
reproduce_barplot_participants()
}
\value{
an object of class ggplot containing the bar plot for participants'
demographics data
}
\description{
Reproduce bar plot for participants' demographics data
}
\examples{
reproduce_barplot_participants()
}
