% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_boxplot_COMPLETENESS}
\alias{reproduce_boxplot_COMPLETENESS}
\title{Reproduce box plot for Completeness}
\usage{
reproduce_boxplot_COMPLETENESS()
}
\value{
an object of class ggplot containing the box plot of Completeness
}
\description{
Reproduce box plot for Completeness
}
\examples{
reproduce_boxplot_COMPLETENESS()
}
