% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_boxplot_CORRECTNESS}
\alias{reproduce_boxplot_CORRECTNESS}
\title{Reproduce box plot for Correctness}
\usage{
reproduce_boxplot_CORRECTNESS()
}
\value{
an object of class ggplot containing the box plot of Correctness
}
\description{
Reproduce box plot for Correctness
}
\examples{
reproduce_boxplot_CORRECTNESS()
}
