% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_confidence_intervals_CORRECTNESS}
\alias{reproduce_confidence_intervals_CORRECTNESS}
\title{Reproduce the confidence intervals for estimates of fixed effects for CORRECTNESS}
\usage{
reproduce_confidence_intervals_CORRECTNESS()
}
\value{
a data frame with parameter name, estimate of fixed effect, standard error and
95% confidence interval limits
}
\description{
Reproduce the confidence intervals for estimates of fixed effects for CORRECTNESS
}
\examples{
reproduce_confidence_intervals_CORRECTNESS()
}
