% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{reproduce_descriptive_stats_COMPLETENESS}
\alias{reproduce_descriptive_stats_COMPLETENESS}
\title{Reproduce Descriptive Statistics for Completeness}
\usage{
reproduce_descriptive_stats_COMPLETENESS()
}
\value{
a dataframe containing the descriptive statistics for Completeness
}
\description{
Reproduce Descriptive Statistics for Completeness
}
\examples{
reproduce_descriptive_stats_COMPLETENESS()
}
