% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{reproduce_descriptive_stats_CORRECTNESS}
\alias{reproduce_descriptive_stats_CORRECTNESS}
\title{Reproduce Descriptive Statistics for Correctness}
\usage{
reproduce_descriptive_stats_CORRECTNESS()
}
\value{
a dataframe containing the descriptive statistics for Correctness
}
\description{
Reproduce Descriptive Statistics for Correctness
}
\examples{
reproduce_descriptive_stats_CORRECTNESS()
}
