% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_emeans_effect_size_COMPLETENESS}
\alias{reproduce_emeans_effect_size_COMPLETENESS}
\title{Reproduce effect sizes based on estimated marginal means}
\usage{
reproduce_emeans_effect_size_COMPLETENESS()
}
\value{
a data frame containing estimated marginal means for Completeness computed for each task
and task description granularity, the difference in estimated marginal means between the
coarser-grained and finer-grained levels and the effect size computed as the percentage increase
in the estimated marginal means
}
\description{
Reproduce effect sizes based on estimated marginal means
}
\examples{
reproduce_emeans_effect_size_COMPLETENESS()
}
