% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_emmeans_COMPLETENESS}
\alias{reproduce_emmeans_COMPLETENESS}
\title{Reproduce estimated marginal means for Completeness}
\usage{
reproduce_emmeans_COMPLETENESS()
}
\value{
a data frame containing estimated marginal means for Completeness computed for each task
and task description granularity
}
\description{
Reproduce estimated marginal means for Completeness
}
\examples{
reproduce_emmeans_COMPLETENESS()
}
