% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_emmeans_CORRECTNESS}
\alias{reproduce_emmeans_CORRECTNESS}
\title{Reproduce estimated marginal means for Corretness}
\usage{
reproduce_emmeans_CORRECTNESS()
}
\value{
a data frame containing estimated marginal means for Correctness computed for each task
and task description granularity
}
\description{
Reproduce estimated marginal means for Corretness
}
\examples{
reproduce_emmeans_CORRECTNESS()
}
