% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_emmeans_profileplot_COMPLETENESS}
\alias{reproduce_emmeans_profileplot_COMPLETENESS}
\title{Reproduce the profile plots of estimated marginal means for Completeness}
\usage{
reproduce_emmeans_profileplot_COMPLETENESS()
}
\value{
a ggplot object containing the profile plot of estimated marginal means for Completeness
}
\description{
Reproduce the profile plots of estimated marginal means for Completeness
}
\examples{
reproduce_emmeans_profileplot_COMPLETENESS()
}
