% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_emmeans_profileplot_CORRECTNESS}
\alias{reproduce_emmeans_profileplot_CORRECTNESS}
\title{Reproduce the profile plots of estimated marginal means for Correctness}
\usage{
reproduce_emmeans_profileplot_CORRECTNESS()
}
\value{
a ggplot object containing the profile plot of estimated marginal means for Correctness
}
\description{
Reproduce the profile plots of estimated marginal means for Correctness
}
\examples{
reproduce_emmeans_profileplot_CORRECTNESS()
}
