% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_estimates_COMPLETENESS}
\alias{reproduce_estimates_COMPLETENESS}
\title{Reproduce the estimates for fixed effects for Completeness}
\usage{
reproduce_estimates_COMPLETENESS()
}
\value{
a data frame which contains the estimates for fixed effects Task and Task Description Granularity
obtained with the the linear mixed-effects model fit for Completeness
}
\description{
Reproduce the estimates for fixed effects for Completeness
}
\examples{
reproduce_estimates_COMPLETENESS()
}
