% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_estimates_CORRECTNESS}
\alias{reproduce_estimates_CORRECTNESS}
\title{Reproduce the estimates of fixed effects, standard errors and confidence intervals for Correctness}
\usage{
reproduce_estimates_CORRECTNESS()
}
\value{
a data frame which contains the estimates of fixed effects Task and Task Description Granularity
obtained with the the linear mixed-effects model fit for Correctness
}
\description{
Reproduce the estimates of fixed effects, standard errors and confidence intervals for Correctness
}
\examples{
reproduce_estimates_CORRECTNESS()
}
