% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_hypothesis_testing_COMPLETENESS}
\alias{reproduce_hypothesis_testing_COMPLETENESS}
\title{Reproduce the hypothesis tests of fixed effects for Completeness}
\usage{
reproduce_hypothesis_testing_COMPLETENESS()
}
\value{
a data frame with numerator degrees of freedom, denominator degrees of freedom,
F-values, and P-values for Wald tests for the fixed effects in the model for Completeness
}
\description{
Reproduce the hypothesis tests of fixed effects for Completeness
}
\examples{
reproduce_hypothesis_testing_COMPLETENESS()
}
