% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_hypothesis_testing_CORRECTNESS}
\alias{reproduce_hypothesis_testing_CORRECTNESS}
\title{Reproduce the hypothesis tests of fixed effects for Correctness}
\usage{
reproduce_hypothesis_testing_CORRECTNESS()
}
\value{
a data frame with numerator degrees of freedom, denominator degrees of freedom,
F-values, and P-values for Wald tests for the fixed effects in the model
}
\description{
Reproduce the hypothesis tests of fixed effects for Correctness
}
\examples{
reproduce_hypothesis_testing_CORRECTNESS()
}
