% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_COMPLETENESS_6}
\alias{reproduce_lmm_COMPLETENESS_6}
\title{Reproduce alternative lmm model 6 for Completeness}
\usage{
reproduce_lmm_COMPLETENESS_6()
}
\value{
an object of class lme which represents the Model 6 for Completeness
}
\description{
Reproduce alternative lmm model 6 for Completeness
}
\examples{
reproduce_lmm_COMPLETENESS_6()
}
