% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_COMPLETENESS_8}
\alias{reproduce_lmm_COMPLETENESS_8}
\title{Reproduce alternative lmm model 8 for Completeness}
\usage{
reproduce_lmm_COMPLETENESS_8()
}
\value{
an object of class lme which represents the Model 8 for Completeness
}
\description{
Reproduce alternative lmm model 8 for Completeness
}
\examples{
reproduce_lmm_COMPLETENESS_8()
}
