% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_CORRECTNESS}
\alias{reproduce_lmm_CORRECTNESS}
\title{Reproduce the LMM model for Correctness}
\usage{
reproduce_lmm_CORRECTNESS()
}
\value{
an object of class nlme::lme which represents the linear mixed-effects model fit
for CORRECTNESS ~ TASK_GRA + TASK + (1|ParticipantID)
}
\description{
Reproduce the LMM model for Correctness
}
\examples{
reproduce_lmm_CORRECTNESS()
}
