% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_CORRECTNESS_1}
\alias{reproduce_lmm_CORRECTNESS_1}
\title{Reproduce alternative lmm model 1 for Correctness}
\usage{
reproduce_lmm_CORRECTNESS_1()
}
\value{
an object of class lme which represents the Model 1 for Correctness
}
\description{
Reproduce alternative lmm model 1 for Correctness
}
\examples{
reproduce_lmm_CORRECTNESS_1()
}
