% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_CORRECTNESS_4}
\alias{reproduce_lmm_CORRECTNESS_4}
\title{Reproduce alternative lmm model 4 for Correctness}
\usage{
reproduce_lmm_CORRECTNESS_4()
}
\value{
an object of class lme which represents the Model 4 for Correctness
}
\description{
Reproduce alternative lmm model 4 for Correctness
}
\examples{
reproduce_lmm_CORRECTNESS_4()
}
