% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_lmm_CORRECTNESS_7}
\alias{reproduce_lmm_CORRECTNESS_7}
\title{Reproduce alternative lmm model 7 for Correctness}
\usage{
reproduce_lmm_CORRECTNESS_7()
}
\value{
an object of class lme which represents the Model 7 for Correctness
}
\description{
Reproduce alternative lmm model 7 for Correctness
}
\examples{
reproduce_lmm_CORRECTNESS_7()
}
