% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_funcions.R
\name{reproduce_model_fit_statistics_CORRECTNESS}
\alias{reproduce_model_fit_statistics_CORRECTNESS}
\title{Reproduce the model fit statistics for the linear-mixed model which represents
the linear mixed-effects model fit for CORRECTNESS}
\usage{
reproduce_model_fit_statistics_CORRECTNESS()
}
\value{
a dataframe containing marginal R-squared (R2m), conditional R-squared (R2c),
Akaike's Information Criterion (AIC), Schwarz's Bayesian criterion (BIC), and
log-likelihood values
}
\description{
Reproduce the model fit statistics for the linear-mixed model which represents
the linear mixed-effects model fit for CORRECTNESS
}
\examples{
reproduce_model_fit_statistics_CORRECTNESS()
}
