% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_violin_plot_COMPLETENESS}
\alias{reproduce_violin_plot_COMPLETENESS}
\title{Reproduce violin plot for Completeness}
\usage{
reproduce_violin_plot_COMPLETENESS()
}
\value{
an object of class ggplot containing the violin plot of Completeness
}
\description{
Reproduce violin plot for Completeness
}
\examples{
reproduce_violin_plot_COMPLETENESS()
}
