% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reproduce_violin_plot_CORRECTNESS}
\alias{reproduce_violin_plot_CORRECTNESS}
\title{Reproduce violin plot for Correctness}
\usage{
reproduce_violin_plot_CORRECTNESS()
}
\value{
an object of class ggplot
}
\description{
Reproduce violin plot for Correctness
}
\examples{
reproduce_violin_plot_CORRECTNESS()
}
