% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducerTaskGra-package.R
\docType{package}
\name{reproducerTaskGra-package}
\alias{reproducerTaskGra}
\alias{reproducerTaskGra-package}
\title{reproducerTaskGra: Reproduce Task Description Granularity Study Results}
\description{
Includes functions for reproducing results of data analysis in Task Description Granularity Study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/M3SOulu/reproducerTaskGra}
  \item Report bugs at \url{https://github.com/M3SOulu/reproducerTaskGra/issues}
}

}
\keyword{internal}
